requirements.txt
Pillow==8.0.1
simplejson==3.16.0